+++ CHANGES.rst
@@ -118,7 +118,7 @@ Changelog
 - Add the ability to add and edit documents in a popup when using datagrid widget with doclinks [djay]
 - Enable caching on value return by selectionlist field [quang]
 - Extend datagrid widget for doclink to select and display document [quang]
-
+- Remove getSelectionList from doclink [quang]
 
 2.0b4 (2016-05-27)
 ------------------